feat(project-stucture): fixed imports of types
