+++ src/components/Header/components/index.ts
@@ -1,2 +1,2 @@
-export { default as ThemeSwitcher } from './ThemeSwitcher/ThemeSwitcher';
-export { default as LanguageSwitcher } from './LanguageSwitcher/LanguageSwitcher';
+export { default as ThemeSwitcher } from './ThemeSwitcher';
+export { default as LanguageSwitcher } from './LanguageSwitcher';
