+++ src/pages/RegisterPage/components/RegisterStep1Form/types.ts
@@ -1,0 +1,4 @@
+export interface RegisterStep1FormProps {
+  onNext: () => Promise<void>;
+  isSubmitting: boolean;
+}
