+++ src/pages/RegisterPage/components/RegisterStep2Form/types.ts
@@ -1,0 +1,4 @@
+export interface RegisterStep2FormProps {
+  onBack: () => void;
+  isSubmitting: boolean;
+}
